refactor(store): move store and add vue shims
