+++ src/pimlify-vue/src/main.ts
@@ -1,18 +1,19 @@
 import Vue from 'vue';
-import App from "./App.vue";
+import App from "./components/App.vue";
 
 import Vuetify from 'vuetify';
-import Vuex from 'vuex';
 
 import 'vuetify/dist/vuetify.css'
 
 import './vuetify-fix.css'
 
+import store from './store'
+
 Vue.use(Vuetify);
-Vue.use(Vuex);
 
 new Vue({
   el: '#app',
+  store,
   render: h => h(App)
 });
