+++ src/pimlify-vue/src/store/index.ts
@@ -1,0 +1,10 @@
+import Vue from "vue";
+import Vuex from "vuex";
+
+import State from "./state" 
+
+Vue.use(Vuex);
+
+export default new Vuex.Store<State>({
+  state: new State()
+});
