+++ src/pimlify-vue/src/store/state.ts
@@ -1,0 +1,3 @@
+export default class State{
+    public message: string = "Hello World!";
+}
